Time: 3 ms (73.14%), Space: 42.8 MB (85.15%) - LeetHub

+++ 0125-valid-palindrome/0125-valid-palindrome.java
@@ -6,8 +6,8 @@ class Solution {
         s = s.toLowerCase();
 
         while (i < j) {
-            int iChar = (int) s.charAt(i);
-            int jChar = (int) s.charAt(j);
+            int iChar = s.charAt(i);
+            int jChar = s.charAt(j);
             if (!Character.isLetterOrDigit(iChar)) {
                 i++;
             } else if (!Character.isLetterOrDigit(jChar)) {
